electron setup
