+++ .gitignore
@@ -1,4 +1,5 @@
 build/
+dist/
 coverage/
 node_modules/
 tmp/
