+++ packages/electron-app/electron-webpack.json
@@ -1,0 +1,5 @@
+{
+  "renderer": {
+    "sourceDirectory": null
+  }
+}
